move exception folder into common foler
