+++ src/main/java/seedu/typists/common/exception/ExceedRangeException.java
@@ -1,4 +1,4 @@
-package seedu.typists.exception;
+package seedu.typists.common.exception;
 
 public class ExceedRangeException extends Throwable {
 }
