+++ src/main/java/seedu/typists/common/exception/FaultyInputException.java
@@ -1,4 +1,4 @@
-package seedu.typists.exception;
+package seedu.typists.common.exception;
 
 public class FaultyInputException extends Exception {
     public FaultyInputException(String errorMessage) {
