+++ src/main/java/seedu/typists/common/exception/IncompleteCommandException.java
@@ -1,4 +1,4 @@
-package seedu.typists.exception;
+package seedu.typists.common.exception;
 
 public class IncompleteCommandException extends Exception {
     //thrown when the command enter is missing key pointns
