+++ src/main/java/seedu/typists/common/exception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package seedu.typists.exception;
+package seedu.typists.common.exception;
 
 public class InvalidCommandException extends Exception {
     public InvalidCommandException(String errorMessage) {
