+++ src/test/java/seedu/StringParserTest.java
@@ -1,7 +1,7 @@
 package seedu;
 
 import org.junit.jupiter.api.Test;
-import seedu.typists.exception.InvalidStringInputException;
+import seedu.typists.common.exception.InvalidStringInputException;
 import seedu.typists.parser.StringParser;
 
 import java.util.ArrayList;
